refactor lb_libraries
