+++ contracts/liquidity_book/lb_factory/Cargo.toml
@@ -18,22 +18,14 @@ default = []
 backtraces = ["cosmwasm-std/backtraces"]
 
 [dependencies]
-shade-protocol = { version = "0.1.0", path = "../../../packages/shade_protocol", features = [
-  "liquidity_book","storage_plus","lb-libraries","swap",
-] }
-cosmwasm-std = { package = "secret-cosmwasm-std", version = "=1.1.11", features = ["iterator"] }
-cosmwasm-storage = { package = "secret-cosmwasm-storage", version = "=1.1.11", features = ["iterator"] }
-cosmwasm-schema = "1.5"
-# schemars = { version = "0.8" }
+shade-protocol = { version = "0.1.0", path = "../../../packages/shade_protocol", features = ["liquidity_book_impl"] }
 serde = { version = "1.0" }
-# serde-json-wasm = { version = "0.5"}
 thiserror = { version = "1.0" }
-secret-storage-plus = { git = "https://github.com/securesecrets/secret-plus-utils", version = "0.13.4", optional = true }
 ethnum = { version = "1.5" }
 
 [dev-dependencies]
 anyhow = "1"
-# cw-multi-test = { version = "0.16.0", default-features = false }
+cosmwasm-std = { package = "secret-cosmwasm-std", version = "1.1.11" }
 
 [[bin]]
 name = "schema"
