+++ contracts/liquidity_book/lb_token/Cargo.toml
@@ -32,9 +32,7 @@ secret-toolkit = { version = "0.10.0", default-features = false, features = [
   "storage", "viewing-key", "permit", "serialization", 
   #"crypto", "utils", 
 ] }
-shade-protocol = { version = "0.1.0", path = "../../../packages/shade_protocol", features = [
-  "liquidity_book","lb-libraries", "snip20"
-] }
+shade-protocol = { version = "0.1.0", path = "../../../packages/shade_protocol", features = ["liquidity_book_impl"] }
 
 #base64 = "0.21.2"
 #rand = { version = "0.8.5", default-features = false }
