+++ contracts/liquidity_book/router/Cargo.toml
@@ -17,22 +17,13 @@ crate-type = ["cdylib", "rlib"]
 
 [features]
 default = []
-# for quicker tests, cargo test --lib
-# for more explicit tests, cargo test --features=backtraces
-
-
 
 [dependencies]
+shade-protocol = { version = "0.1.0", path = "../../../packages/shade_protocol", features = ["liquidity_book_impl"] }
 snafu = { version = "0.7.1" }
 serde = { version = "1.0.114", default-features = false, features = [
   "derive",
   "alloc"
-] }
-cosmwasm-schema = "1.5.0"
-cosmwasm-std = { package = "secret-cosmwasm-std", version = "=1.1.11", features = ["iterator"] }
-cosmwasm-storage = { package = "secret-cosmwasm-storage", version = "=1.1.11", features = ["iterator"] }
-shade-protocol = { version = "0.1.0", path = "../../../packages/shade_protocol", features = [
-  "liquidity_book","storage_plus","lb-libraries","swap",
 ] }
 
 [dev-dependencies]
