+++ contracts/liquidity_book/tests/Cargo.toml
@@ -18,21 +18,16 @@ default = []
 backtraces = ["cosmwasm-std/backtraces"]
 
 [dependencies]
-shade-protocol = { version = "0.1.0", path = "../../../packages/shade_protocol", features = [
-  "liquidity_book","storage_plus","lb-libraries","swap"
-] }
-cosmwasm-std = { package = "secret-cosmwasm-std", version = "=1.1.11", features = ["iterator"] }
-cosmwasm-storage = { package = "secret-cosmwasm-storage", version = "=1.1.11", features = ["iterator"] }
-cosmwasm-schema = "1.5"
+shade-protocol = { version = "0.1.0", path = "../../../packages/shade_protocol", features = ["liquidity_book_impl"] }
 serde = { version = "1.0" }
 serde-json-wasm = { version = "1.0"}
-secret-storage-plus = { git = "https://github.com/securesecrets/secret-plus-utils", version = "0.13.4", optional = true }
 ethnum = { version = "1.5" }
 rand = "0.8"
 thiserror = { version = "1.0" }
 
 [dev-dependencies]
 anyhow = "1"
+cosmwasm-std = { package = "secret-cosmwasm-std", version = "1.1.11" }
 shade-multi-test = { path = "../../../packages/multi_test", features = [ 
   "admin",
   "lb_pair",
