+++ packages/shade_protocol/Cargo.toml
@@ -36,7 +36,6 @@ storage = ["dao-utils"]
 storage_plus = ["storage", "dep:secret-storage-plus", "chrono"]
 query_auth_lib = ["dep:query-authentication"]
 multi-test = ["dep:secret-multi-test", "dep:anyhow", "interface"]
-lb-libraries = ["swap", "snip20", "interface", "liquidity_book"]
 
 # Implementing new contracts:
 # Contracts packages are divided by two different features, the interface and the implementation
@@ -50,7 +49,7 @@ interface = ["utils", "errors"]
 # #[cfg(feature = "contract_name")]
 # pub mod contract_name;
 
-# For the implementation you need to import this in your feature and prepend the feature name with `_impl`
+# For the implementation you need to import this in your feature and append the feature name with `_impl`
 implementation = ["storage_plus", "storage", "admin"]
 # Like so
 # contract_name_impl = ["contract_name", "implementation"]
@@ -88,8 +87,8 @@ dao              = ["interface", "cosmwasm-std/staking"]
 admin            = ["interface"]
 peg_stability    = ["sky-utils", "adapter"]
 snip20_migration = []
+swap             = ["interface", "snip20"]
 liquidity_book   = ["interface", "swap"]
-swap             = ["snip20"]
 
 chrono = ["dep:chrono"]
 
@@ -101,12 +100,12 @@ mock = []
 
 # Protocol Implementation Contracts
 # Used in internal smart contracts
-governance-impl = ["implementation", "governance"]
-snip20-impl     = ["snip20", "query_auth_impl"]
-query_auth_impl = ["implementation", "query_auth", "dep:base64"]
-sky-utils       = ["implementation", "sky"]
-admin_impl      = ["implementation", "admin"]
-# liquidity_book_impl = ["liquidity_book", "implementation", "swap"]
+governance-impl     = ["implementation", "governance"]
+snip20-impl         = ["snip20", "query_auth_impl"]
+query_auth_impl     = ["implementation", "query_auth", "dep:base64"]
+sky-utils           = ["implementation", "sky"]
+admin_impl          = ["implementation", "admin"]
+liquidity_book_impl = ["implementation", "liquidity_book", "snip20", "swap"]
 
 # for quicker tests, cargo test --lib
 # for more explicit tests, cargo test --features=backtraces
