+++ packages/shade_protocol/src/contract_interfaces/liquidity_book/lb_libraries/lb_token/state_structs.rs
