+++ packages/shade_protocol/src/contract_interfaces/liquidity_book/lb_libraries/math/encoded_sample.rs
@@ -3,7 +3,7 @@
 //!
 //! Helper library used for setting and decoding parts of encoded Bytes32.
 
-use crate::utils::liquidity_book::types::Bytes32;
+use crate::lb_libraries::types::Bytes32;
 use cosmwasm_schema::cw_serde;
 use ethnum::U256;
