+++ packages/shade_protocol/src/contract_interfaces/liquidity_book/lb_libraries/math/u128x128_math.rs
@@ -6,7 +6,7 @@
 
 use ethnum::{I256, U256};
 
-use crate::utils::liquidity_book::constants::*;
+use crate::lb_libraries::constants::*;
 
 use super::bit_math::BitMath;
