+++ packages/shade_protocol/src/contract_interfaces/liquidity_book/lb_libraries/math/u24.rs
