+++ packages/shade_protocol/src/contract_interfaces/liquidity_book/mod.rs
@@ -1,3 +1,6 @@
 pub mod lb_factory;
 pub mod lb_pair;
 pub mod lb_token;
+
+#[cfg(all(feature = "liquidity_book_impl", feature = "swap"))]
+pub mod lb_libraries;
